+++ R/run-source.r
@@ -128,19 +128,19 @@ find_gist <- function(id, filename) {
 
   if (!is.null(filename)) {
     if (!is.character(filename) || length(filename) > 1 || !grepl("\\.[rR]$", filename)) {
-      stop("'filename' must be NULL, or a single filename ending in .R/.r")
+      stop("'filename' must be NULL, or a single filename ending in .R/.r", call. = FALSE)
     }
 
     which <- match(tolower(filename), tolower(names(r_files)))
     if (is.na(which)) {
-      stop("You have speficied a file that is not in this gist.")
+      stop("'", filename, "' not found in this gist", call. = FALSE)
     }
 
   } else {
     if (length(r_files) > 1) {
-      warning("Multiple R files in gist, using first.")
-      which <- 1
+      warning("Multiple R files in gist, using first", call. = FALSE)
     }
+    which <- 1
   }
 
   r_files[[which]]$raw_url