+++ NEWS.md
@@ -1,5 +1,8 @@
 # devtools 1.13.3.9000
 
+* `source_gist()` works once more when there is only a single file
+  in the gist (#1266).
+
 # devtools 1.13.2
 Workaround a regression in Rcpp::compileAttributes.
 Add trimws implementation for R 3.1 support.